requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
pandas
Pillow
gdown